+++ transformations/examples/Readme.md
@@ -1,2 +1,5 @@
 ## repository-problems:
 - dagrejs is a fork of dagre, which is not maintained. graphlib is the underlying library from the original maintainer
+
+## resources:
+- using the typescript compiler API directly: https://github.com/microsoft/TypeScript/wiki/Using-the-Compiler-API
