--- transformations/examples/classes.json
@@ -1,1 +1,0 @@
-[]
