--- transformations/examples/src/docGenerator.ts
@@ -1,127 +1,0 @@
-import * as ts from "typescript";
-import * as fs from "fs";
-
-//this script could help generate documentation
-//copied from here: https://github.com/Microsoft/TypeScript/wiki/Using-the-Compiler-API
-//ts-node looks more easy to use though
-
-interface DocEntry {
-  name?: string;
-  fileName?: string;
-  documentation?: string;
-  type?: string;
-  constructors?: DocEntry[];
-  parameters?: DocEntry[];
-  returnType?: string;
-}
-
-/** Generate documentation for all classes in a set of .ts files */
-function generateDocumentation(
-  fileNames: string[],
-  options: ts.CompilerOptions
-): void {
-  console.log("fileNames: " + fileNames);
-  // Build a program using the set of root file names in fileNames
-  let program = ts.createProgram(fileNames, options);
-
-  // Get the checker, we will use it to find more about classes
-  let checker = program.getTypeChecker();
-  let output: DocEntry[] = [];
-
-  // Visit every sourceFile in the program
-  for (const sourceFile of program.getSourceFiles()) {
-	ts.forEachChild(sourceFile, visit);
-    if (!sourceFile.isDeclarationFile) {
-      // Walk the tree to search for classes
-      //ts.forEachChild(sourceFile, visit);
-    } else {
-      console.log("sourceFile.moduleName: " + sourceFile.moduleName);
-      console.log("sourceFile.fileName: " + sourceFile.fileName);
-      console.log("sourceFile.referencedFiles: " + sourceFile.referencedFiles);
-    }
-  }
-
-  // print out the doc
-  fs.writeFileSync("classes.json", JSON.stringify(output, undefined, 4));
-
-  return;
-
-  /** visit nodes finding exported classes */
-  function visit(node: ts.Node) {
-    // Only consider exported nodes
-    if (!isNodeExported(node)) {
-      return;
-    }
-    if (ts.isInterfaceDeclaration(node) && !ts.isImportClause(node)) {
-      console.log("is interface " + node.name.getText());
-    }
-    //
-    if (ts.isClassDeclaration(node) && node.name) {
-      // This is a top level class, get its symbol
-      let symbol = checker.getSymbolAtLocation(node.name);
-      if (symbol) {
-        output.push(serializeClass(symbol));
-      }
-      // No need to walk any further, class expressions/inner declarations
-      // cannot be exported
-    } else if (ts.isModuleDeclaration(node)) {
-      // This is a namespace, visit its children
-      ts.forEachChild(node, visit);
-    }
-    //
-  }
-
-  /** Serialize a symbol into a json object */
-  function serializeSymbol(symbol: ts.Symbol): DocEntry {
-    return {
-      name: symbol.getName(),
-      documentation: ts.displayPartsToString(
-        symbol.getDocumentationComment(checker)
-      ),
-      type: checker.typeToString(
-        checker.getTypeOfSymbolAtLocation(symbol, symbol.valueDeclaration!)
-      ),
-    };
-  }
-
-  /** Serialize a class symbol information */
-  function serializeClass(symbol: ts.Symbol) {
-    let details = serializeSymbol(symbol);
-
-    // Get the construct signatures
-    let constructorType = checker.getTypeOfSymbolAtLocation(
-      symbol,
-      symbol.valueDeclaration!
-    );
-    details.constructors = constructorType
-      .getConstructSignatures()
-      .map(serializeSignature);
-    return details;
-  }
-
-  /** Serialize a signature (call or construct) */
-  function serializeSignature(signature: ts.Signature) {
-    return {
-      parameters: signature.parameters.map(serializeSymbol),
-      returnType: checker.typeToString(signature.getReturnType()),
-      documentation: ts.displayPartsToString(
-        signature.getDocumentationComment(checker)
-      ),
-    };
-  }
-
-  /** True if this is visible outside this file, false otherwise */
-  function isNodeExported(node: ts.Node): boolean {
-    return (
-      (ts.getCombinedModifierFlags(node as ts.Declaration) &
-        ts.ModifierFlags.Export) !==
-        0 ||
-      (!!node.parent && node.parent.kind === ts.SyntaxKind.SourceFile)
-    );
-  }
-}
-
-generateDocumentation(process.argv.slice(2), {
-  target: ts.ScriptTarget.ES5,
-  module: ts.ModuleKind.CommonJS,
-});
