+++ transformations/examples/src/interfaces.ts
@@ -1,0 +1,7 @@
+/**
+ * a return type for matching algorithms running over typescript programs. Contains the exports that match, as well as 'others'
+ */
+export interface ExportsMatcherSummary {
+  matchingExports: string[];
+  otherExports: string[];
+}
