--- transformations/examples/testdata/Header.d.ts
@@ -1,7 +1,0 @@
-export interface HeaderProps {
-    user?: {};
-    onLogin: () => void;
-    onLogout: () => void;
-    onCreateAccount: () => void;
-}
-export declare const Header: React.FC<HeaderProps>;
