--- transformations/examples/testdata/mui-import-file.ts
@@ -1,2 +1,0 @@
-
-import muiLib from "@mui/material";
